Исправление ошибок postman

+++ backend/api/views.py
@@ -1,22 +1,21 @@
-from django.conf import settings
-from django_filters.rest_framework import DjangoFilterBackend
-from django.shortcuts import get_object_or_404
-from rest_framework import filters, mixins, status, viewsets
-from rest_framework.decorators import action
-from rest_framework.permissions import (AllowAny, IsAuthenticated,
-                                        IsAuthenticatedOrReadOnly)
-from rest_framework.response import Response
-
 from core.filters import IngredientFilter, RecipeFilter
 from core.paginations import ApiPagination
-from core.permissions import IsAdminOrReadOnly
+from core.permissions import IsAdminOrReadOnly, IsAuthenticatedOr401
+from django.conf import settings
+from django.shortcuts import get_object_or_404
+from django_filters.rest_framework import DjangoFilterBackend
 from recipes.models import Favorite, Ingredient, Recipe, Tag
+from rest_framework import filters, mixins, status, viewsets
+from rest_framework.decorators import action
+from rest_framework.permissions import AllowAny
+from rest_framework.response import Response
 from shortlink.models import UrlShort
 from users.models import FoodgramUser, Subscription
+
 from .serializers import (FavoriteSerializer, FoodgramUserSerializer,
-                          IngredientSerializer, RecipeReadSerializer,
-                          RecipeWriteSerializer, TagSerializer,
-                          SubscriptionSerializer)
+                          IngredientSerializer, PasswordChangeSerializer,
+                          RecipeReadSerializer, RecipeWriteSerializer,
+                          SubscriptionSerializer, TagSerializer)
 
 
 class FoodgramUserViewSet(viewsets.ModelViewSet):
@@ -24,27 +23,23 @@ class FoodgramUserViewSet(viewsets.ModelViewSet):
 
     queryset = FoodgramUser.objects.all()
     serializer_class = FoodgramUserSerializer
-    permission_classes = (IsAuthenticatedOrReadOnly, IsAdminOrReadOnly)
+    permission_classes = (IsAuthenticatedOr401, IsAdminOrReadOnly)
     filter_backends = (filters.SearchFilter,)
     search_fields = ("username",)
-    http_method_names = ("get", "post", "patch", "delete")
+    http_method_names = ("get", "post", "put", "delete")
     lookup_field = "id"
     pagination_class = ApiPagination
 
     @action(
         detail=False,
         methods=("get", "patch"),
-        permission_classes=(IsAuthenticated,),
+        permission_classes=(IsAuthenticatedOr401,),
     )
     def me(self, request):
         """Получение и изменение информации о текущем пользователе."""
 
         user = request.user
-        if user.is_anonymous:
-            return Response(
-                {"detail": "Авторизуйтесь для доступа в профиль."},
-                status=status.HTTP_401_UNAUTHORIZED
-            )
+
         if request.method == "GET":
             serializer = self.get_serializer(user)
             return Response(serializer.data)
@@ -61,32 +56,67 @@ class FoodgramUserViewSet(viewsets.ModelViewSet):
 
     @action(
         detail=False,
-        methods=("get", "post", "delete", "patch"),
-        permission_classes=(IsAuthenticated,),
+        methods=("put", "delete"),
+        permission_classes=(IsAuthenticatedOr401,),
         url_path='me/avatar'
     )
     def avatar(self, request):
-        """Загрузка аватара текущего пользователя."""
+        """Изменение и удаление аватара текущего пользователя."""
 
         user = request.user
-        file = request.FILES.get('avatar')
-        if not file:
-            return Response({"detail": "Нет файла."},
+
+        if request.method == "PUT":
+            data = request.data.get('avatar')
+            if not data:
+                return Response({"detail": "Аватара нет."},
+                                status=status.HTTP_400_BAD_REQUEST
+                                )
+
+            serializer = self.get_serializer(
+                user, data={'avatar': data}, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                base_url = getattr(settings,
+                                   'DOMAIN_URL',
+                                   'http://localhost:8000'
+                                   )
+                return Response(
+                    {"avatar": f"{base_url}/{user.avatar}"},
+                    status=status.HTTP_200_OK
+                )
+            return Response(serializer.errors,
                             status=status.HTTP_400_BAD_REQUEST
                             )
 
-        user.avatar = file
-        user.save()
-        return Response({"detail": "Аватар загружен."},
-                        status=status.HTTP_200_OK
-                        )
+        elif request.method == "DELETE":
+            user.avatar.delete(save=True)
+            return Response({"detail": "Аватар удалён."},
+                            status=status.HTTP_204_NO_CONTENT
+                            )
+
+    @action(
+        detail=False,
+        methods=['post'],
+        permission_classes=(IsAuthenticatedOr401,),
+        url_path='set_password'
+    )
+    def set_password(self, request):
+        """Изменение пароля текущего пользователя."""
+
+        user = request.user
+        serializer = PasswordChangeSerializer(
+            data=request.data, context={'request': request})
+
+        if serializer.is_valid():
+            user.set_password(serializer.validated_data['new_password'])
+            user.save()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def get_permissions(self):
         if self.action == 'create':
             self.permission_classes = (AllowAny,)
-        else:
-            self.permission_classes = (
-                IsAuthenticatedOrReadOnly, IsAdminOrReadOnly)
         return super(FoodgramUserViewSet, self).get_permissions()
 
     def get_serializer_context(self):
@@ -173,10 +203,13 @@ class RecipeViewSet(viewsets.ModelViewSet):
         short_url = f"{base_url}/s/{url_short.short_url}"
         return Response({"short-link": short_url})
 
-    @action(detail=True, methods=['post', 'delete'], url_path='favorite')
+    @action(detail=True,
+            methods=['post', 'delete'],
+            url_path='favorite')
     def favorite(self, request, *args, **kwargs):
         if not request.user.is_authenticated:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         user = request.user
         recipe = get_object_or_404(Recipe, id=self.kwargs.get('pk'))
         serializer_context = {
@@ -217,50 +250,47 @@ class SubscriptionViewSet(mixins.CreateModelMixin,
                           viewsets.GenericViewSet):
     """Вьюсет для управления подписками на авторов."""
     pagination_class = ApiPagination
+    permission_classes = (IsAuthenticatedOr401,)
 
     @action(detail=True,
-            methods=['post'],
-            url_path='subscribe')
+            methods=['post', 'delete'],
+            url_path='subscribe',
+            permission_classes=(IsAuthenticatedOr401,),)
     def subscribe(self, request, id=None):
-        """Добавляем подписку."""
-
-        author = get_object_or_404(FoodgramUser, id=id)
-        user = request.user
-        if Subscription.objects.filter(user=user, author=author).exists():
-            return Response(
-                {"detail": "Такая подписка уже существует."},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-        if author == user:
-            return Response(
-                {"detail": "Подписаться на себя нельзя(хоть и хочется)."},
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        subscription = Subscription.objects.create(user=user, author=author)
-        serializer = SubscriptionSerializer(
-            subscription, context={'request': request})
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @action(detail=True,
-            methods=['delete'],
-            url_path='subscribe')
-    def unsubscribe(self, request, id=None):
-        """Удаляем подписку."""
+        """Управление подписками."""
 
         author = get_object_or_404(FoodgramUser, id=id)
         user = request.user
 
-        subscription = Subscription.objects.filter(
-            user=user, author=author).first()
-        if subscription:
-            subscription.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.method == 'POST':
+            if Subscription.objects.filter(user=user, author=author).exists():
+                return Response(
+                    {"detail": "Такая подписка уже существует."},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            if author == user:
+                return Response(
+                    {"detail": "Подписаться на себя нельзя(хоть и хочется)."},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
 
-        return Response(
-            {"detail": "Подписка не найдена."},
-            status=status.HTTP_400_BAD_REQUEST
-        )
+            subscription = Subscription.objects.create(
+                user=user, author=author)
+            serializer = SubscriptionSerializer(
+                subscription, context={'request': request})
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        if request.method == 'DELETE':
+            subscription = Subscription.objects.filter(
+                user=user, author=author)
+            if subscription:
+                subscription.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+
+            return Response(
+                {"detail": "Подписка не найдена."},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
     @action(detail=False,
             methods=['get'],
